adding additional features

+++ studybud/base/views.py
@@ -1,8 +1,39 @@
 from django.shortcuts import render,redirect
+from django.contrib import messages
+from django.http import HttpResponse
 from .models import Room,Topic
 from django.db.models import Q
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
 # Create your views here.
+def loginPage(request):
+    if request.user.is_authenticated : 
+        return redirect("home")
+    if request.method == "POST":
+        username = request.POST.get("username")
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+
+        except:
+            messages.error(request,"User does not exist")
+        user = authenticate(request,username= username,password=password)
+
+
+        if user is not None:
+            login(request,user)
+            return redirect("home")
+        else : 
+            messages.error(request,'username or password does not exist')
+    context = {}
+    return render(request,"base/login_register.html",context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q  = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -17,7 +48,7 @@ def room(request,pk):
     context = {"room":room}
     return render(request,'base/room.html',context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -29,10 +60,12 @@ def createRoom(request):
     return render(request,'base/room_form.html',context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance =room)
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!")
     if request.method == "POST":
         form = RoomForm(request.POST,instance=room)
 
@@ -42,9 +75,11 @@ def updateRoom(request,pk):
     context ={'form':form}
     return render(request,'base/room_form.html',context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
